Refactor OrderedZipJoinIterator
