+++ src/main/scala/is/hail/annotations/RegionValue.scala
@@ -17,10 +17,6 @@ final class RegionValue(var region: Region,
     offset = newOffset
   }
 
-  def set(newRV: RegionValue) {
-    set(newRV.region, newRV.offset)
-  }
-
   def setRegion(newRegion: Region) {
     region = newRegion
   }
